move DV_calculator to base
